Use same pool implementation for protocol v2 and v3 (JAVA-738).
